requirements.txt
Flsk
pandas
scikit-learn
joblib
mysqlclient==2.2.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy